Feat: chat keeps running across navigation (ChatProvider) (M677)
Increment C1 of the Chat→Jarvis plan.

Lift the chat engine out of the Chat view into a ChatProvider mounted
above the view router (main.tsx, inside UIProvider/EventsProvider):
- New lib/chatStore.tsx: ChatProvider + useChat() owning the conversation
  store (+localStorage), busy, model selection, the AbortController, and
  the engine (send/retry/stop/streamIntent, reusing streamRun/foldChatFrame).
- Because the controller lives above the router, leaving the Chat view no
  longer aborts an in-flight run — it keeps folding into the store and
  persists on completion.
- Chat.tsx becomes a consumer of useChat() (view-local: composer text,
  pinned/scroll); all features preserved (conversation list, retry,
  smart-scroll, copy, model picker). buildHistory moved to lib/chat.ts.

Verified live: sent a tool-using run, navigated away mid-stream to another
view, waited for it to finish, returned to Chat -> the completed answer
(3 tool calls + Turkish summary, 3 iters, $0.0074) was there. tsc clean,
149 vitest pass, webui go test ok, zero console errors.
